requirements.txt
fastapi>=0.111
uvicorn>=0.29
boto3>=1.34
python-dotenv>=1.0
